requirements.txt
fastapi
uvicorn
beanie
python-dotenv
python-jose
bcrypt
pymongo[srv]
pydantic[email]